feature: add benches feature to make bench tests optional
