+++ core/src/block.rs
@@ -788,7 +788,6 @@ fn is_for_avm(block: &mut OpenBlock, tx: &SignedTransaction) -> bool {
 }
 
 #[inline]
-#[cfg(not(feature = "slow-blocks"))]
 fn push_transactions(
     block: &mut OpenBlock,
     transactions: &[SignedTransaction],
@@ -829,37 +828,6 @@ fn push_transactions(
 
     trace!(target: "vm", "push transactions done");
 
-    Ok(())
-}
-
-#[cfg(feature = "slow-blocks")]
-fn push_transactions(
-    block: &mut OpenBlock,
-    transactions: &[SignedTransaction],
-) -> Result<(), Error>
-{
-    use std::time;
-
-    let slow_tx = option_env!("SLOW_TX_DURATION")
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(100);
-    for t in transactions {
-        let hash = t.hash();
-        let start = time::Instant::now();
-        block.push_transaction(t.clone(), None, false)?;
-        let took = start.elapsed();
-        let took_ms = took.as_secs() * 1000 + took.subsec_nanos() as u64 / 1000000;
-        if took > time::Duration::from_millis(slow_tx) {
-            warn!(
-                target: "tx",
-                "Heavy ({} ms) transaction in block {:?}: {:?}",
-                took_ms,
-                block.header().number(),
-                hash
-            );
-        }
-        debug!(target: "tx", "Transaction {:?} took: {} ms", hash, took_ms);
-    }
     Ok(())
 }
