+++ core/src/lib.rs
@@ -21,7 +21,6 @@
  ******************************************************************************/
 
 #![warn(unused_extern_crates)]
-#![cfg_attr(feature = "benches", feature(test))]
 extern crate bloomchain;
 extern crate byteorder;
 extern crate crossbeam;
