+++ keystore/key/Cargo.toml
@@ -9,3 +9,6 @@ crypto = { path = "../../util/crypto" }
 aion-types = { path = "../../util/aion-types"}
 rand = "0.4"
 rustc-hex = "1.0"
+
+[features]
+benches = []
