+++ util/crypto/Cargo.toml
@@ -11,3 +11,6 @@ gcc = "^0.3"
 rustc-serialize = "^0.3"
 rand = "*"
 libc = "*"
+
+[features]
+benches = []
