+++ util/ethbloom/Cargo.toml
@@ -26,3 +26,4 @@ default = ["std", "heapsizeof", "serialize", "fixed-hash/libc"]
 std = ["fixed-hash/std"]
 heapsizeof = ["fixed-hash/heapsizeof"]
 serialize = ["std", "ethereum-types-serialize", "serde"]
+benches = []
