+++ util/fixed_point/Cargo.toml
@@ -11,3 +11,6 @@ lazy_static = "1.3.0"
 
 [dev-dependencies]
 rand = "0.4"
+
+[features]
+benches = []
