+++ util/triehash/tests/lib.rs
@@ -22,42 +22,9 @@
 
 #![warn(unused_extern_crates)]
 
-extern crate trie_standardmap;
 extern crate triehash;
-extern crate aion_types;
-extern crate blake2b;
 
-use std::time::Instant;
-use aion_types::H256;
-use blake2b::blake2b;
-use trie_standardmap::{Alphabet, ValueMode, StandardMap};
 use triehash::{trie_root, shared_prefix_len, hex_prefix_encode};
-
-fn random_word(alphabet: &[u8], min_count: usize, diff_count: usize, seed: &mut H256) -> Vec<u8> {
-    assert!(min_count + diff_count <= 32);
-    *seed = blake2b(&seed);
-    let r = min_count + (seed[31] as usize % (diff_count + 1));
-    let mut ret: Vec<u8> = Vec::with_capacity(r);
-    for i in 0..r {
-        ret.push(alphabet[seed[i] as usize % alphabet.len()]);
-    }
-    ret
-}
-
-fn random_bytes(min_count: usize, diff_count: usize, seed: &mut H256) -> Vec<u8> {
-    assert!(min_count + diff_count <= 32);
-    *seed = blake2b(&seed);
-    let r = min_count + (seed[31] as usize % (diff_count + 1));
-    seed[0..r].to_vec()
-}
-
-fn random_value(seed: &mut H256) -> Vec<u8> {
-    *seed = blake2b(&seed);
-    match seed[0] % 2 {
-        1 => vec![seed[31]; 1],
-        _ => seed.to_vec(),
-    }
-}
 
 #[test]
 fn test_hex_prefix_encode() {
@@ -139,160 +106,4 @@ fn test_shared_prefix3() {
     let a = vec![1, 2, 3, 4, 5, 6];
     let b = vec![1, 2, 3, 4, 5, 6];
     assert_eq!(shared_prefix_len(&a, &b), 6);
-}
-
-#[test]
-fn benchtest_triehash_insertions_32_mir_1k() {
-    let st = StandardMap {
-        alphabet: Alphabet::All,
-        min_key: 32,
-        journal_key: 0,
-        value_mode: ValueMode::Mirror,
-        count: 1000,
-    };
-    let d = st.make();
-
-    let count = 1000;
-    let time = Instant::now();
-
-    let mut result = H256::default();
-    for _ in 0..count {
-        result = trie_root(d.clone()).clone();
-    }
-
-    assert!(result.0.len() != 0);
-
-    let took = time.elapsed();
-    println!(
-        "[benchtest_triehash_insertions_32_mir_1k] triehash insertions 32 mirror 1k (ns/call): {}",
-        (took.as_secs() * 1000_000_000 + took.subsec_nanos() as u64) / count
-    );
-}
-
-#[test]
-fn benchtest_triehash_insertions_32_ran_1k() {
-    let st = StandardMap {
-        alphabet: Alphabet::All,
-        min_key: 32,
-        journal_key: 0,
-        value_mode: ValueMode::Random,
-        count: 1000,
-    };
-    let d = st.make();
-
-    let count = 1000;
-    let time = Instant::now();
-
-    let mut result = H256::default();
-    for _ in 0..count {
-        result = trie_root(d.clone()).clone();
-    }
-
-    assert!(result.0.len() != 0);
-
-    let took = time.elapsed();
-    println!(
-        "[benchtest_triehash_insertions_32_ran_1k] triehash insertions 32 random 1k (ns/call): {}",
-        (took.as_secs() * 1000_000_000 + took.subsec_nanos() as u64) / count
-    );
-}
-
-#[test]
-fn benchtest_triehash_insertions_six_high() {
-    let mut d: Vec<(Vec<u8>, Vec<u8>)> = Vec::new();
-    let mut seed = H256::new();
-    for _ in 0..1000 {
-        let k = random_bytes(6, 0, &mut seed);
-        let v = random_value(&mut seed);
-        d.push((k, v))
-    }
-
-    let count = 1000;
-    let time = Instant::now();
-
-    for _ in 0..count {
-        trie_root(d.clone());
-    }
-
-    let took = time.elapsed();
-    println!(
-        "[benchtest_triehash_insertions_six_high] triehash insertions six high (ns/call): {}",
-        (took.as_secs() * 1000_000_000 + took.subsec_nanos() as u64) / count
-    );
-}
-
-#[test]
-fn benchtest_triehash_insertions_six_mid() {
-    let alphabet = b"@QWERTYUIOPASDFGHJKLZXCVBNM[/]^_";
-    let mut d: Vec<(Vec<u8>, Vec<u8>)> = Vec::new();
-    let mut seed = H256::new();
-    for _ in 0..1000 {
-        let k = random_word(alphabet, 6, 0, &mut seed);
-        let v = random_value(&mut seed);
-        d.push((k, v))
-    }
-
-    let count = 1000;
-    let time = Instant::now();
-
-    for _ in 0..count {
-        trie_root(d.clone());
-    }
-
-    let took = time.elapsed();
-    println!(
-        "[benchtest_triehash_insertions_six_mid] triehash insertions six mid (ns/call): {}",
-        (took.as_secs() * 1000_000_000 + took.subsec_nanos() as u64) / count
-    );
-}
-
-#[test]
-fn benchtest_triehash_insertions_random_mid() {
-    let alphabet = b"@QWERTYUIOPASDFGHJKLZXCVBNM[/]^_";
-    let mut d: Vec<(Vec<u8>, Vec<u8>)> = Vec::new();
-    let mut seed = H256::new();
-    for _ in 0..1000 {
-        let k = random_word(alphabet, 1, 5, &mut seed);
-        let v = random_value(&mut seed);
-        d.push((k, v))
-    }
-
-    let count = 1000;
-    let time = Instant::now();
-
-    for _ in 0..count {
-        trie_root(d.clone());
-    }
-
-    let took = time.elapsed();
-    println!(
-        "[benchtest_triehash_insertions_random_mid] triehash insertions six random mid (ns/call): \
-         {}",
-        (took.as_secs() * 1000_000_000 + took.subsec_nanos() as u64) / count
-    );
-}
-
-#[test]
-fn benchtest_triehash_insertions_six_low() {
-    let alphabet = b"abcdef";
-    let mut d: Vec<(Vec<u8>, Vec<u8>)> = Vec::new();
-    let mut seed = H256::new();
-    for _ in 0..1000 {
-        let k = random_word(alphabet, 6, 0, &mut seed);
-        let v = random_value(&mut seed);
-        d.push((k, v))
-    }
-
-    let count = 1000;
-    let time = Instant::now();
-
-    for _ in 0..count {
-        trie_root(d.clone());
-    }
-
-    let took = time.elapsed();
-    println!(
-        "[benchtest_triehash_insertions_six_low] triehash insertions six low (ns/call): {}",
-        (took.as_secs() * 1000_000_000 + took.subsec_nanos() as u64) / count
-    );
 }
